pyproject.toml
[project]
name = "fastapiproject"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "MyNextWeekend", email = "jinhu007@outlook.com" }
]
dependencies = [
    "fastapi[standard]>=0.115.4",
    "uvicorn>=0.32.0",
    "gitpython>=3.1.43",
    "docker>=7.1.0",
    "sqlmodel>=0.0.22",
    "pymysql>=1.1.1",
    "kafka-python>=2.0.2",
    "pydantic-settings>=2.6.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/fastapiproject"]

[dependency-groups]
dev = [
    "pytest>=8.3.3",
]


# uv 配置下载源地址
#[tool.uv]
#index-url = "https://pypi.tuna.tsinghua.edu.cn/simple"

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG001", # unused arguments in functions
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
    "B904",  # Allow raising exceptions without from e, for HTTPException
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true
